+++ changelog.md
@@ -5,6 +5,7 @@
 - Improvement: Introduced the local selftest
 - Improvement: Added CLI parameter to the SSTableExaminer
 - Bugfix: Init table splitter only in distributed mode
+- Bugfix: In a major compact, not all tables are processed
 
 ### Version 0.2.0 (Beta-3) - 09.12.2016
 - New Feature: Introduced paging (results > memory could be requested)